backend/requirements.txt
fastapi
uvicorn[standard]
pydantic>=2.0
python-dotenv
google-generativeai
openai
anthropic
Pillow
pyyaml
# chromadb  # optional: install for vector-based RAG retrieval
